+++ packages/cli/__tests__/zosjobs/__integration__/view/__scripts__/all-spool-content/job_syntax_invalid_parm.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+zowe zos-jobs view all-spool-content "JOB123" blah --host fakehost --user fakeuser --password fakepass
+exit $?